scatterapi/requirements.txt
django 
gunicorn 
djangorestframework
psycopg2-binary
django-channels
celery
daphne
django-rest-knox
redis
channels-redis
django-cors-headers==3.1.0
Pillow